r Renamer localStorage-utils til browserStorage-utils for enklere løskobling fra localStorage
